refactor: trim warehouse sql helper surface
